Dev

See merge request !12

+++ updatetool/make.sh
@@ -4,16 +4,17 @@ Usage()
 {
     echo "============================================="
     echo "使用方式:"
-    echo "\t./make 型号 地区\n"
+    echo "\t./make 型号 地区 说明内容\n"
     echo "\t设备类型，1：I型集中器，2:II型集中器，3：III型专变"
-    echo "\tZheJiang(II型),ShanDong(II型),HuNan(I型)"
+    echo "\tZheJiang(II型),ShanDong(II型),HuNan(I型),GW(国网送检)"
+    echo "\t说明内容"
     echo "============================================="
     exit 1
 }
 
 ParaCheck()
 {
-    if [ "$#" -ne "2" ]; then
+    if [ "$#" -ne "3" ]; then
         Usage
     fi
 
@@ -21,10 +22,11 @@ ParaCheck()
         Usage
     fi
 
-    if [ $2 != "ZheJiang" ] && [ $2 != "HuNan" ] && [ $2 != "ShanDong" ]; then
+    if [ $2 != "ZheJiang" ] && [ $2 != "HuNan" ] && [ $2 != "ShanDong" ] && [ $2 != "GW" ]  && [ $2 != "GWFILE" ]; then
          Usage
     fi
-
+    
+#    echo $1 $2 $3  
     echo "创建必要条件..."
     if [ ! -d app ]; then
         mkdir app
@@ -41,18 +43,29 @@ Clean()
 CopyNew()
 {
     echo "复制最新的程序与库..."
-    cp ../bin_arm/cj* ./app/
-    cp ../bin_arm/*.so ./app/
-    cp ../config/* ./app/
+    if [ $1 = "GWFILE" ]; then
+        cp ../bin_arm/cjmain ./app/
+        cp ../bin_arm/libBase.so ./app/
+    else
+        cp ../bin_arm/cj* ./app/
+        cp ../bin_arm/*.so ./app/
+        cp ../config/* ./app/
+        rm ./app/systema_2.cfg
+        if [ $1 = "GW" ]; then
+            cp ../config/systema_2.cfg ./app/systema.cfg
+        fi
+    fi
 }
 
 UpdateLocation()
 {
-    rm app/device.cfg
-    echo "[device]" >> app/device.cfg
-    echo "device="$1 >> app/device.cfg
-    echo "zone="$2 >> app/device.cfg
-    echo "[end]" >> app/device.cfg
+    if [ $2 != "GWFILE" ]; then
+        rm app/device.cfg
+        echo "[device]" >> app/device.cfg
+        echo "device="$1 >> app/device.cfg
+        echo "zone="$2 >> app/device.cfg
+        echo "[end]" >> app/device.cfg
+    fi
 }
 
 Package()
@@ -96,11 +109,11 @@ Composer()
 {
     echo "生成集合包..."
     #$path = 'history/$1.$2.$(date +%Y%m%d%H)'
-    mkdir -p history/$1.$2.$(date +%Y%m%d%H)
-    cp -R QCheck history/$1.$2.$(date +%Y%m%d%H)
-    cp app.tar.gz history/$1.$2.$(date +%Y%m%d%H)
-    cp -R cjgwn history/$1.$2.$(date +%Y%m%d%H)
-    cp update.sh history/$1.$2.$(date +%Y%m%d%H)
+    mkdir -p history/$1.$2.$(date +%Y%m%d%H).$3
+    cp -R QCheck history/$1.$2.$(date +%Y%m%d%H).$3
+    cp app.tar.gz history/$1.$2.$(date +%Y%m%d%H).$3
+    cp -R cjgwn history/$1.$2.$(date +%Y%m%d%H).$3
+    cp update.sh history/$1.$2.$(date +%Y%m%d%H).$3
 }
 
 Post_Clean()
@@ -112,15 +125,15 @@ Post_Clean()
 
 main()
 {
-    ParaCheck $1 $2
+    ParaCheck $1 $2 $3
     Clean
-    CopyNew
+    CopyNew $2
     UpdateLocation $1 $2
     Package
     CreateUSB
     Tools $1 $2
-    Composer $1 $2
+    Composer $1 $2 $3
     Post_Clean
 }
 
-main $1 $2
+main $1 $2 $3
